backend/requirements-py312.txt
# Alternative requirements for Python 3.12 (recommended for stability)
# Use this if you encounter build issues with Python 3.13

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.2.0
python-multipart==0.0.9
pypdf==4.0.1
openai==1.12.0
pinecone-client==2.2.4

